+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.3.5
 - Updated to conform with CloudHarvestCoreTasks 0.7.0
 - Updated the index definitions in the `harvest.yaml` file
+- Flask HTTP log now routed to the main logger
 
 ## 0.3.4
 - Added try when aggregating results in `tasks/get_task_status()` which prevented the status from being returned